Revert previous commit
