+++ src/app/mainview/mainview.component.ts
@@ -29,6 +29,8 @@ import { NgForm, FormGroup, FormBuilder, FormControl, Validators } from '@angula
 import { Predictioninterval } from 'app/shared/interfaces/predictioninterval';
 import { URLSearchParams } from '@angular/http';
 import { Citation } from 'app/shared/interfaces/citation';
+import { connectableObservableDescriptor } from 'rxjs/internal/observable/ConnectableObservable';
+
 
 declare var MathJax: {
     Hub: { Queue };
@@ -102,6 +104,7 @@ export class MainviewComponent implements OnInit, OnDestroy {
     public itemBeingEdited;
     public uniqueRegRegions;
     public editScen;
+    public cloneScen;
     public paramsNeeded;
     public getBounds: boolean;
     private modalSubscript;
@@ -175,6 +178,7 @@ export class MainviewComponent implements OnInit, OnDestroy {
             if (region) { this.showRegion = true; }
             window.scrollTo(0, 0);
         });
+
         this.loggedInRole = localStorage.getItem('loggedInRole');
         MathJax.Hub.Queue(['Typeset', MathJax.Hub, 'mathJaxLoad']); // preload mathjax
         // this is based on a behaviorSubject, so it gets an initial notification of [].
@@ -1206,11 +1210,22 @@ export class MainviewComponent implements OnInit, OnDestroy {
         popupWin.document.close();
     }
 
-    /////////////////////// Add/Edit/Delete Scenarios Section ///////////////////////////
-    public showAddScenarioModal() {
+    /////////////////////// Clone Scenarios Section ///////////////////////////
+    public showCloneScenarioModal() {
         this._nssService.setAddScenarioModal(true);
     }
 
+    newCloneScenario(cloneScen){
+        this._nssService.changeItem(cloneScen);
+    }
+
+    public cloneRowClicked(statisticGroupID, r, rr) {
+        this.cloneScen={r,rr,statisticGroupID};
+        this.newCloneScenario(this.cloneScen);
+        this.showCloneScenarioModal();
+    }
+
+    /////////////////////// Edit Scenarios Section ///////////////////////////
     public editRegScenario() {
         this.editRegionScenario = true;
         MathJax.Hub.Queue(['Typeset', MathJax.Hub, 'mathJax1']); // render equations into Mathjax
@@ -1221,8 +1236,81 @@ export class MainviewComponent implements OnInit, OnDestroy {
         if (this.itemBeingEdited) { this.CancelEditRowClicked(); }
     }
 
+    public editRowClicked(item, rrIndex, sgIndex, idx?) {
+        if (this.itemBeingEdited && this.itemBeingEdited.isEditing && this.tempData && this.itemBeingEdited.name !== item.name) {
+            this.CancelEditRowClicked();
+        } // if another item was being edited, cancel that
+        this.tempData = JSON.parse(JSON.stringify(item)); // make a copy in case they cancel
+        idx >= 0 ? this.editIdx = idx : this.editIdx = null;
+        this.editRRindex = rrIndex; this.editSGIndex = sgIndex; // setting indices because the cancel function wasn't overwriting things
+        this.itemBeingEdited = item;
+        item.isEditing = true;
+        if (item.equation) {this.showMathjax(item); }
+        if (idx === undefined) {
+            // do we need the citation IDs?
+            const rrIdx = this.regressionRegions.findIndex(rr => rr.id === item.id);
+            this.scenarios[this.editSGIndex].regressionRegions[this.editRRindex].citationID = this.regressionRegions[rrIdx].citationID;
+        }
+    }
+
+    public CancelEditRowClicked() {
+        // reset item if cancelling editing
+        if (this.editIdx === null) { // if regression region
+            this.scenarios[this.editSGIndex].regressionRegions[this.editRRindex] = this.tempData;
+        } else if (this.itemBeingEdited.citationURL) { // if citation
+            this.citations[this.editIdx] = this.tempData;
+        } else if (this.itemBeingEdited.limits) { // if parameter
+            this.scenarios[this.editSGIndex].regressionRegions[this.editRRindex].parameters[this.editIdx] = this.tempData;
+        } else if (this.itemBeingEdited.equation) { // if regression
+            this.scenarios[this.editSGIndex].regressionRegions[this.editRRindex].regressions[this.editIdx] = this.tempData;
+            const equ = document.getElementById('mathjaxEq' + this.itemBeingEdited.id);
+            equ.style.visibility = 'hidden';
+            MathJax.Hub.Queue(['Typeset', MathJax.Hub, 'mathJax1']);
+        }
+    }
+
+    /////////////////////// Delete Scenarios Section ///////////////////////////
+    deleteRegression(sgID, rrID, rID) {
+        const check = confirm('Are you sure you want to delete this Regression?');
+        if (check) {
+            const sParams: URLSearchParams = new URLSearchParams();
+            sParams.set('statisticgroupID', sgID);
+            sParams.set('regressionregionID', rrID);
+            sParams.set('regressiontypeID', rID);
+            this._settingsService.deleteEntity('', this.configSettings.scenariosURL, sParams).subscribe(result => {
+                this._nssService.setSelectedRegion(this.selectedRegion);
+                if (result.headers) { this._nssService.outputWimMessages(result); }
+            }, error => {
+                if (error.headers) {this._nssService.outputWimMessages(error);
+                } else { this._nssService.handleError(error); }
+            });
+        }
+    }
+
+    deleteRegRegion(rrID) {
+        const check = confirm('Are you sure you want to delete this Regression Region?');
+        if (check) {
+            this._settingsService.deleteEntity(rrID, this.configSettings.regRegionURL).subscribe(result => {
+                this._nssService.setSelectedRegion(this.selectedRegion);
+                if (result.headers) { this._nssService.outputWimMessages(result); }
+            }, error => {
+                if (error.headers) {this._nssService.outputWimMessages(error);
+                } else { this._nssService.handleError(error); }
+            });
+        }
+    }
+
+    
+    /////////////////////// Citations Section ///////////////////////////
     public showManageCitationsModal() {
         this._nssService.setManageCitationsModal(true);
+    }
+
+    public getCitations() {
+        this._settingsService.getEntities(this.configSettings.citationURL)
+            .subscribe(res => {
+                this.citations = res;
+            });
     }
 
     public saveCitation(c) {
@@ -1269,6 +1357,11 @@ export class MainviewComponent implements OnInit, OnDestroy {
                 }
             );
         }
+    }
+    
+    /////////////////////// Add Scenarios Section ///////////////////////////
+    public showAddScenarioModal() {
+        this._nssService.setAddScenarioModal(true);
     }
 
     public saveParameter(p, rrIndex, sgIndex) {
@@ -1355,46 +1448,6 @@ export class MainviewComponent implements OnInit, OnDestroy {
             });
     }
 
-    public getCitations() {
-        this._settingsService.getEntities(this.configSettings.citationURL)
-            .subscribe(res => {
-                this.citations = res;
-            });
-    }
-
-    public editRowClicked(item, rrIndex, sgIndex, idx?) {
-        if (this.itemBeingEdited && this.itemBeingEdited.isEditing && this.tempData && this.itemBeingEdited.name !== item.name) {
-            this.CancelEditRowClicked();
-        } // if another item was being edited, cancel that
-        this.tempData = JSON.parse(JSON.stringify(item)); // make a copy in case they cancel
-        idx >= 0 ? this.editIdx = idx : this.editIdx = null;
-        this.editRRindex = rrIndex; this.editSGIndex = sgIndex; // setting indices because the cancel function wasn't overwriting things
-        this.itemBeingEdited = item;
-        item.isEditing = true;
-        if (item.equation) {this.showMathjax(item); }
-        if (idx === undefined) {
-            // do we need the citation IDs?
-            const rrIdx = this.regressionRegions.findIndex(rr => rr.id === item.id);
-            this.scenarios[this.editSGIndex].regressionRegions[this.editRRindex].citationID = this.regressionRegions[rrIdx].citationID;
-        }
-    }
-
-    public CancelEditRowClicked() {
-        // reset item if cancelling editing
-        if (this.editIdx === null) { // if regression region
-            this.scenarios[this.editSGIndex].regressionRegions[this.editRRindex] = this.tempData;
-        } else if (this.itemBeingEdited.citationURL) { // if citation
-            this.citations[this.editIdx] = this.tempData;
-        } else if (this.itemBeingEdited.limits) { // if parameter
-            this.scenarios[this.editSGIndex].regressionRegions[this.editRRindex].parameters[this.editIdx] = this.tempData;
-        } else if (this.itemBeingEdited.equation) { // if regression
-            this.scenarios[this.editSGIndex].regressionRegions[this.editRRindex].regressions[this.editIdx] = this.tempData;
-            const equ = document.getElementById('mathjaxEq' + this.itemBeingEdited.id);
-            equ.style.visibility = 'hidden';
-            MathJax.Hub.Queue(['Typeset', MathJax.Hub, 'mathJax1']);
-        }
-    }
-
     public addError(errors) {
         // if user adds error, push empty object to array
         errors.push({});
@@ -1403,36 +1456,6 @@ export class MainviewComponent implements OnInit, OnDestroy {
     compareObjects(Obj1, Obj2) {
         // used to make sure the existing options are showing in selects
         return Obj1 && Obj2 ? Obj1.id === Obj2.id : Obj1 === Obj2;
-    }
-
-    deleteRegression(sgID, rrID, rID) {
-        const check = confirm('Are you sure you want to delete this Regression?');
-        if (check) {
-            const sParams: URLSearchParams = new URLSearchParams();
-            sParams.set('statisticgroupID', sgID);
-            sParams.set('regressionregionID', rrID);
-            sParams.set('regressiontypeID', rID);
-            this._settingsService.deleteEntity('', this.configSettings.scenariosURL, sParams).subscribe(result => {
-                this._nssService.setSelectedRegion(this.selectedRegion);
-                if (result.headers) { this._nssService.outputWimMessages(result); }
-            }, error => {
-                if (error.headers) {this._nssService.outputWimMessages(error);
-                } else { this._nssService.handleError(error); }
-            });
-        }
-    }
-
-    deleteRegRegion(rrID) {
-        const check = confirm('Are you sure you want to delete this Regression Region?');
-        if (check) {
-            this._settingsService.deleteEntity(rrID, this.configSettings.regRegionURL).subscribe(result => {
-                this._nssService.setSelectedRegion(this.selectedRegion);
-                if (result.headers) { this._nssService.outputWimMessages(result); }
-            }, error => {
-                if (error.headers) {this._nssService.outputWimMessages(error);
-                } else { this._nssService.handleError(error); }
-            });
-        }
     }
 
     showInputModal() {
@@ -1613,7 +1636,32 @@ export class MainviewComponent implements OnInit, OnDestroy {
         this.modalRef.close();
     }
 
-    
+    private createNewRegression() {
+        const regionID = this.newRegRegForm.value.state;
+        this._settingsService
+            .postEntity(this.newRegRegForm.value, this.configSettings.regionURL + regionID + '/' + this.configSettings.regRegionURL)
+            .subscribe(
+                (response) => {
+                    response.isEditing = false;
+                    if (!response.headers) {this._toasterService.pop('info', 'Info', 'Regression region was added');
+                    } else {this._settingsService.outputWimMessages(response); }
+                    if (this.addCitation) { // if user elected to add a citation, send that through
+                        this.createNewCitation(response);
+                    } else {
+                        this.cancelCreateRegression();
+                        this._nssService.setSelectedRegion(this.selectedRegion);
+                    }
+                    if (this.uploadPolygon) { // if user elected to upload a polygon, send that through
+                        this.uploadNewPolygon(response);
+                    } else {
+                        this.cancelCreateRegression();
+                        this._nssService.setSelectedRegion(this.selectedRegion);
+                    }
+                }, error => {
+                    if (this._settingsService.outputWimMessages(error)) {return; }
+                    this._toasterService.pop('error', 'Error creating Regression Region', error._body.message || error.statusText); }
+            );
+    }
 
     public createNewCitation(rr) {
         // add new citation
@@ -1637,7 +1685,6 @@ export class MainviewComponent implements OnInit, OnDestroy {
     public uploadNewPolygon(rr) {
         // Upload new polygon
         // TODO: Send polygon to services
-        console.log("Polygon will be sent to services.")
     }
 
     private getDismissReason(reason: any): string {
