+++ src/app/shared/components/addregressionregion/addregressionregion.component.html
@@ -10,7 +10,7 @@
   <div class="modal-body">
       <h3 class="modal-title" *ngIf="addRegReg">New Regression Region</h3>
       <h3 class="modal-title" *ngIf="!addRegReg">New Citation</h3>
-      <form *ngIf="addRegReg" [formGroup]="newRegRegForm" styl="margin-bottom: 10px;">
+      <form *ngIf="addRegReg" [formGroup]="newRegRegForm" style="margin-bottom: 10px;">
           <!-- Regression Type Name* -->
           <div class="form-group required" [ngClass]="{'form-invalid': !newRegRegForm.get('name').valid && newRegRegForm.get('name').dirty}">
               <label class="req" for="name">Name:</label>
